cleanup C++20: Adjust enum values (Use 1U << x for unsigned values; reduce type size)

+++ include/jau/functional.hpp
@@ -338,7 +338,7 @@ namespace jau {
          *
          * @see @ref function_overview "Function Overview"
          */
-        enum class target_type : uint32_t {
+        enum class target_type : uint16_t {
             /** Denotes a func::null_target_t */
             null = 0,
             /** Denotes a func::member_target_t */
